Update signin.php
Instead of using if conditions to display error messages, the messages are stored in JavaScript and added to the page dynamically using the DOM. This separates the PHP logic from the HTML and makes the code more maintainable.

The error messages are now inside a div with an id of result. This makes it easier to reference and manipulate the messages using JavaScript.