Update user-view.php
Code has been refactored to reduce duplicated code and make it easier to understand and maintain. 

The $user_image and $profile_name variables are used to store the values that would otherwise be repeated multiple times. 
The $is_new_member variable is used to determine if the user is a new member or not, which is then used to conditionally display the "New Member" text.

Note: please name the CSS classes properly and watch out for spelling mistakes, e.g. 'skill_bobbies_'